Include pip-tools in development dependencies
This is mostly to fix the version of pip-tools. The format of the
requirements.txt that pip-compile generates is known to change every
once in a while. It's good to know with what version of pip-tools the
current file is generated. Now it's generated with version 6.12.3 (and
as you see, the format changed a little).