Serialize fields that are relations to another model correctly
If a model that is serialized has a field that is a relation to another
model, that related model's (or manager's) `serialize` method needs to
be called.

Tested this by serializing also the User model that Profile model has a
one-to-one relation to. The "user" field is now serialized twice, once
via its `serialize` method and once with the special "accessor".