Enable configuring a user provider function
It's required that a User model can be obtained from the model instance
specified by the `GDPR_API_MODEL` setting. By default the model
instance's `user` attribute is tried. If that won't work, it's now
possible to configure a function with the `GDPR_API_USER_PROVIDER`
setting. That function can do anything that is needed to obtain the
user.