Enable configuring the GDPR API URL pattern
The URL pattern string can be given with the `GDPR_API_URL_PATTERN`
setting.

The setting affects the Djano URLConf, which is cached after the first
load. There's a trick in conftest.py that reloads the URLConfs every
time this setting changes. That trick isn't needed during runtime of the
app because settings shouldn't change during runtime.