Accept any version of drf-oidc-auth
It's unknown what's the real reason behind this version limitation.

My guess is it was there because Tunnistamo puts the `amr` claim into
JWTs as a string, while according to specifications it should be an
array of strings. Django-helusers is used for checking the JWTs and it
uses drf-oidc-auth for the heavy lifting. Drf-oidc-auth versions >=1.0.0
demand the `amr` claim to be standards compliant and thus refuse to
accept Tunnistamo JWTs. However, there is now a fix for this in
django-helusers [1], which allows using newer versions of drf-oidc-auth.

[1] https://github.com/City-of-Helsinki/django-helusers/commit/95a7421a51e8d66f59b8bcea240a0e26e57eb35d